Update doc (#195)
* Update docs; add models.md.

* Update vmafossexec.md; move h5py dependency out of mixin.py.

* Make loading svmutil on-demand in quality_runner.py.

* Make loading svmutil on-demand in train_test_model.py.

* Update vmafossexec.md.

* Fix wrapper link.

* Update vmafossexec.md.

* Update libvmaf.md.

* Update VMAF_Python_library.md; move models part to models.md.

* Update VMAF_Python_library.md.

* Update VMAF_Python_library.md.

* Update VMAF_Python_library.md.

* Add references.md and conf_interval.md; re-org docs.

* Misc.

* Update FAQ.md; misc.

* Update models.md.

* Update models.md.

* Update conf_interval.md.

* Add CI plot to conf_interval.md.

* Update image size.

* Update image.

* Update conf_interval.md.

* Update references.md.

* Update references.md.

* Update references.md.

* Update references.md.

* Update references.md.

* Update references.md.

* Misc update of docs.
