+++ resource/doc/datasets.md
@@ -1,7 +1,7 @@
 Datasets
 ===================
 
-We also provide two sample datasets including the video files and the properly formatted dataset files in Python.
+We also provide two sample datasets including the video files and the properly formatted dataset files in Python. They can be used as sample datasets to train and test custom VMAF models.
 
 ### Netflix Public Dataset
